project/requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
diff-match-patch==20121119
Django==2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-import-export==1.0.1
djangorestframework==3.7.7
et-xmlfile==1.0.1
idna==2.6
jdcal==1.4
Markdown==2.6.11
ndg-httpsclient==0.4.3
numpy==1.14.2
odfpy==1.3.6
opencv-python==3.4.3.18
openpyxl==2.5.4
Pillow==5.0.0
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.5.0
pyperclip==1.6.0
pytz==2017.3
PyYAML==3.13
requests==2.19.1
selenium==3.11.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
